refactor: refactor error to a common place
