+++ __mocks__/openstack.ts
@@ -3,6 +3,10 @@ import {
   OpenStackCloudData,
   OpenStackCloudInput,
 } from '@ducks/openstack/cloud/types';
+import {
+  OpenStackProjectData,
+  OpenStackProjectInput,
+} from '@ducks/openstack/project/types';
 
 export const openstackCloudData: OpenStackCloudData = {
   id: 123,
@@ -17,8 +21,24 @@ export const openstackCloudData: OpenStackCloudData = {
   owner_group_id: '3fa85f64-5717-4562-b3fc-2c963f66afa6',
 };
 
+export const openstackProjectData: OpenStackProjectData = {
+  cloud_id: 1,
+  cloud_name: 'string',
+  description: 'string',
+  group_id: '3fa85f64-5717-4562-b3fc-2c963f66afa6',
+  group_name: 'string',
+  id: 1,
+  name: 'myproject',
+  owner_id: '3fa85f64-5717-4562-b3fc-2c963f66afa6',
+  owner_name: 'string',
+};
+
 export const openstackCloudInput: OpenStackCloudInput = {
   ...openstackCloudData,
+};
+
+export const openstackProjectInput: OpenStackProjectInput = {
+  ...openstackProjectData,
 };
 
 export const errorExample = {
