+++ src/store/ducks/lab/types.ts
@@ -6,10 +6,3 @@ export interface Quota {
   num_volumes: number;
   volumes_gb: number;
 }
-
-export interface Error {
-  detail: string;
-  status: number;
-  title: string;
-  type: string;
-}
