+++ src/store/ducks/types.ts
@@ -1,0 +1,6 @@
+export interface Error {
+  detail: string;
+  status: number;
+  title: string;
+  type: string;
+}
